Adds modularity by moving constants and thr reducer function into separate files
