+++ frontend/src/constants/constants.js
@@ -1,0 +1,14 @@
+export const ACTIONS = {
+  FAV_PHOTO_ADDED: 'FAV_PHOTO_ADDED',
+  FAV_PHOTO_REMOVED: 'FAV_PHOTO_REMOVED',
+  SET_PHOTO_DATA: 'SET_PHOTO_DATA',
+  SET_TOPIC_DATA: 'SET_TOPIC_DATA',
+  SELECT_PHOTO: 'SELECT_PHOTO',
+  DISPLAY_PHOTO_DETAILS: 'DISPLAY_PHOTO_DETAILS',
+  CLOSE_PHOTO_DETAILS: 'CLOSE_PHOTO_DETAILS'
+};
+
+export const API_ENDPOINTS = {
+  PHOTOS: '/api/photos',
+  TOPICS: '/api/topics',
+};
