+++ frontend/src/hooks/useApplicationData.js
@@ -1,60 +1,6 @@
 import { useEffect, useReducer } from 'react';
-
-export const ACTIONS = {
-  FAV_PHOTO_ADDED: 'FAV_PHOTO_ADDED',
-  FAV_PHOTO_REMOVED: 'FAV_PHOTO_REMOVED',
-  SET_PHOTO_DATA: 'SET_PHOTO_DATA',
-  SET_TOPIC_DATA: 'SET_TOPIC_DATA',
-  SELECT_PHOTO: 'SELECT_PHOTO',
-  DISPLAY_PHOTO_DETAILS: 'DISPLAY_PHOTO_DETAILS',
-  CLOSE_PHOTO_DETAILS: 'CLOSE_PHOTO_DETAILS'
-};
-
-// Reducer function
-function reducer(state, action) {
-  switch (action.type) {
-    case ACTIONS.FAV_PHOTO_ADDED:
-      return {
-        ...state,
-        likedPhotos: [...state.likedPhotos, action.photoId]
-      };
-    case ACTIONS.FAV_PHOTO_REMOVED:
-      return {
-        ...state,
-        likedPhotos: state.likedPhotos.filter(id => id !== action.photoId)
-      };
-    case ACTIONS.SET_PHOTO_DATA:
-      return {
-        ...state,
-        photos: action.photos // Populate photos with actual data
-      };
-    case ACTIONS.SET_TOPIC_DATA:
-      return {
-        ...state,
-        topics: action.topics // Populate topics with actual data
-      };
-    case ACTIONS.SELECT_PHOTO:
-      return {
-        ...state,
-        selectedPhoto: action.photo
-      };
-    case ACTIONS.DISPLAY_PHOTO_DETAILS:
-      return {
-        ...state,
-        isModalOpen: true
-      };
-
-    case ACTIONS.CLOSE_PHOTO_DETAILS:
-      return {
-        ...state,
-        isModalOpen: false,
-        selectedPhoto: null
-      }
-
-    default:
-      throw new Error(`Unsupported action type: ${action.type}`);
-  }
-};
+import { API_ENDPOINTS, ACTIONS } from '../constants/constants'
+import { reducer } from '../reducers/applicationDataReducer';
 
 // Define the custom hook
 const useApplicationData = () => {
@@ -68,11 +14,11 @@ const useApplicationData = () => {
     topics: []
   });
 
-  // Fetch the photo data and topics data store them in state
+  // Fetch the photo data and topics data and store them in state
   useEffect(() => {
     Promise.all([
-      fetch(`/api/photos`).then(res => res.json()),
-      fetch(`/api/topics`).then( res => res.json()) 
+      fetch(API_ENDPOINTS.PHOTOS).then(res => res.json()),
+      fetch(API_ENDPOINTS.TOPICS).then( res => res.json()) 
     ])
       .then(([photosData, topicsData]) => {
         dispatch({ type: ACTIONS.SET_PHOTO_DATA, photos: photosData });
@@ -83,7 +29,6 @@ const useApplicationData = () => {
       });
   },[]);
 
-  // Function to get similar photos
   const getSimilarPhotos = (selectedPhoto) => {
     if (!selectedPhoto) return [];
     // Extract similar photos from the selected photo object
