+++ frontend/src/reducers/applicationDataReducer.js
@@ -1,0 +1,55 @@
+export const ACTIONS = {
+  FAV_PHOTO_ADDED: 'FAV_PHOTO_ADDED',
+  FAV_PHOTO_REMOVED: 'FAV_PHOTO_REMOVED',
+  SET_PHOTO_DATA: 'SET_PHOTO_DATA',
+  SET_TOPIC_DATA: 'SET_TOPIC_DATA',
+  SELECT_PHOTO: 'SELECT_PHOTO',
+  DISPLAY_PHOTO_DETAILS: 'DISPLAY_PHOTO_DETAILS',
+  CLOSE_PHOTO_DETAILS: 'CLOSE_PHOTO_DETAILS'
+};
+
+// Reducer function
+export function reducer(state, action) {
+  switch (action.type) {
+    case ACTIONS.FAV_PHOTO_ADDED:
+      return {
+        ...state,
+        likedPhotos: [...state.likedPhotos, action.photoId]
+      };
+    case ACTIONS.FAV_PHOTO_REMOVED:
+      return {
+        ...state,
+        likedPhotos: state.likedPhotos.filter(id => id !== action.photoId)
+      };
+    case ACTIONS.SET_PHOTO_DATA:
+      return {
+        ...state,
+        photos: action.photos // Populate photos with actual data
+      };
+    case ACTIONS.SET_TOPIC_DATA:
+      return {
+        ...state,
+        topics: action.topics // Populate topics with actual data
+      };
+    case ACTIONS.SELECT_PHOTO:
+      return {
+        ...state,
+        selectedPhoto: action.photo
+      };
+    case ACTIONS.DISPLAY_PHOTO_DETAILS:
+      return {
+        ...state,
+        isModalOpen: true
+      };
+
+    case ACTIONS.CLOSE_PHOTO_DETAILS:
+      return {
+        ...state,
+        isModalOpen: false,
+        selectedPhoto: null
+      }
+
+    default:
+      throw new Error(`Unsupported action type: ${action.type}`);
+  }
+};
